renamed AVM spider, reworked references to class variables using self and cls in dlink and avm, moved duplicate itemloader method to custom FirmwareSpider superclass
