+++ firmware/custom_spiders.py
@@ -1,11 +1,24 @@
 from abc import ABCMeta
+from typing import Generator
 
 from scrapy import Spider
+from scrapy.loader import ItemLoader
 
 from firmware.custom_requests import FTPFileRequest, FTPListRequest
+from firmware.items import FirmwareItem
 
 
-class FTPSpider(Spider, metaclass=ABCMeta):
+class FirmwareSpider(Spider, metaclass=ABCMeta):
+
+    @staticmethod
+    def item_pipeline(meta_data: dict) -> Generator[FirmwareItem, None, None]:
+        loader = ItemLoader(item=FirmwareItem(), selector=meta_data['file_urls'])
+        for key, value in meta_data.items():
+            loader.add_value(key, value)
+        yield loader.load_item()
+
+
+class FTPSpider(FirmwareSpider, metaclass=ABCMeta):
 
     def start_requests(self):
         for url in self.start_urls:
